+++ pkg/widgets/meshgrid/meshgrid_mouse.go
@@ -18,6 +18,7 @@ func (m *Meshgrid) MouseMoved(event *desktop.MouseEvent) {
 	dy := float64(event.Position.Y - m.lastMouseY)
 	if m.dragging {
 		if event.Button&desktop.MouseButtonPrimary == desktop.MouseButtonPrimary {
+			//m.orbit(dx*rotationScale, -dy*rotationScale)
 			m.rotateMeshgrid(-dy*rotationScale, dx*rotationScale, 0)
 			m.throttledRefresh()
 		} else if event.Button&desktop.MouseButtonSecondary == desktop.MouseButtonSecondary {